+++ cmd/google.go
@@ -5,18 +5,12 @@ import (
 	"fmt"
 
 	kitlog "github.com/go-kit/kit/log"
-	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 
-	"github.com/cycloidio/terracognita/filter"
 	"github.com/cycloidio/terracognita/google"
-	"github.com/cycloidio/terracognita/hcl"
 	"github.com/cycloidio/terracognita/log"
-	"github.com/cycloidio/terracognita/provider"
-	"github.com/cycloidio/terracognita/state"
 	"github.com/cycloidio/terracognita/tag"
-	"github.com/cycloidio/terracognita/writer"
 )
 
 var (
@@ -69,36 +63,9 @@ var (
 				return err
 			}
 
-			f := &filter.Filter{
-				Tags:    tags,
-				Include: include,
-				Exclude: exclude,
-				Targets: targets,
-			}
-
-			var hclW, stateW writer.Writer
-			options, err := getWriterOptions()
+			err = importProvider(ctx, logger, googleP)
 			if err != nil {
 				return err
-			}
-
-			if hclOut != nil {
-				logger.Log("msg", "initializing HCL writer")
-				hclW = hcl.NewWriter(hclOut, googleP, options)
-			}
-
-			if stateOut != nil {
-				logger.Log("msg", "initializing TFState writer")
-				stateW = state.NewWriter(stateOut, options)
-			}
-
-			logger.Log("msg", "importing")
-
-			fmt.Fprintf(logsOut, "Starting Terracognita with version %s\n", Version)
-			logger.Log("msg", "starting terracognita", "version", Version)
-			err = provider.Import(ctx, googleP, hclW, stateW, f, logsOut)
-			if err != nil {
-				return errors.Wrap(err, "could not import from google")
 			}
 
 			return nil